# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API with Swagger documentation.",
        "title": "UnfallAPI",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/unfallBeteiligung": {
            "get": {
                "description": "(46241-0011/46241-0012) Unfallbeteiligte, Hauptverursacher des Unfalls: Deutschland, Jahre, Monate, Geschlecht, Altersgruppen, Art der Verkehrsbeteiligung, Unfallkategorie, Ortslage",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GENESIS-Online (Die Datenbank des Statistischen Bundesamtes)"
                ],
                "summary": "Unfallbeteiligte, Hauptverursacher des Unfalls abrufen",
                "parameters": [
                    {
                        "enum": [
                            "Kraftrad mit Versicherungskennzeichen",
                            "Kraftrad mit amtlichem Kennzeichen",
                            "Elektrokleinstfahrzeuge",
                            "Personenkraftwagen",
                            "Kraftomnibus",
                            "Güterkraftfahrzeug",
                            "Landwirtschaftliche Zugmaschine",
                            "Übrige Kraftfahrzeuge",
                            "Fahrrad ohne Hilfsmotor",
                            "Pedelecs",
                            "Andere Fahrzeuge",
                            "Fußgänger",
                            "Andere Personen",
                            "Insgesamt"
                        ],
                        "type": "string",
                        "description": "Filter nach Art der Verkehrsbeteiligung",
                        "name": "verkehrsart",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "Unfälle mit Personenschaden",
                            "Schwerwiegende Unfälle mit Sachschaden i.e.S",
                            "Sonst. Unfälle unter dem Einfluss berausch. Mittel",
                            "Übrige Sachschadensunfälle",
                            "Insgesamt"
                        ],
                        "type": "string",
                        "description": "Filter nach Unfallkategorie",
                        "name": "unfallkategorie",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter nach Ortslage (innerorts, außerorts (ohne Autobahnen), auf Autobahnen, Insgesamt)",
                        "name": "ortslage",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "männlich",
                            "weiblich",
                            "Ohne Angabe",
                            "Insgesamt"
                        ],
                        "type": "string",
                        "description": "Filter nach Geschlecht",
                        "name": "geschlecht",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "unter 15 Jahre",
                            "15 bis unter 18 Jahre",
                            "18 bis unter 21 Jahre",
                            "21 bis unter 25 Jahre",
                            "25 bis unter 35 Jahre",
                            "35 bis unter 45 Jahre",
                            "45 bis unter 55 Jahre",
                            "55 bis unter 65 Jahre",
                            "65 bis unter 75 Jahre",
                            "75 Jahre und mehr",
                            "Alter unbekannt",
                            "Insgesamt"
                        ],
                        "type": "string",
                        "description": "Filter nach Altersgruppe",
                        "name": "altersgruppe",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "Unfallbeteiligte",
                            "Hauptverursacher des Unfalls"
                        ],
                        "type": "string",
                        "description": "Filter nach Beteiligungsart",
                        "name": "beteiligungsart",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter nach Jahr",
                        "name": "jahr",
                        "in": "query"
                    },
                    {
                        "maximum": 12,
                        "minimum": 0,
                        "type": "integer",
                        "description": "Filter nach Monat (1-12, 0 für Ganzjahresdaten)",
                        "name": "monat",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.UnfallBeteiligung"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid parameter type or range",
                        "schema": {
                            "$ref": "#/definitions/handlers.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Database execution or scanning failure",
                        "schema": {
                            "$ref": "#/definitions/handlers.HTTPError"
                        }
                    }
                }
            }
        },
        "/unfallBeteiligung/jahre": {
            "get": {
                "description": "Gibt alle Jahre zurück, für die Daten vorhanden sind.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GENESIS-Online (Die Datenbank des Statistischen Bundesamtes)"
                ],
                "summary": "Verfügbare Jahre abrufen",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.YearsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Database execution or scanning failure",
                        "schema": {
                            "$ref": "#/definitions/handlers.HTTPError"
                        }
                    }
                }
            }
        },
        "/unfallFehlverhalten": {
            "get": {
                "description": "(46241-0009/46241-0010) Unfallbeteiligte: Deutschland, Jahre, Monate, Art der Verkehrsbeteiligung, Fehlverhalten der Fahrzeugführer und Fußgänger bei Unfällen mit Personenschaden",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GENESIS-Online (Die Datenbank des Statistischen Bundesamtes)"
                ],
                "summary": "Fehlverhalten bei Unfällen mit Personenschaden abrufen",
                "parameters": [
                    {
                        "enum": [
                            "Kraftrad mit Versicherungskennzeichen",
                            "Kraftrad mit amtlichem Kennzeichen",
                            "Elektrokleinstfahrzeuge",
                            "Personenkraftwagen",
                            "Kraftomnibus",
                            "Güterkraftfahrzeug",
                            "Landwirtschaftliche Zugmaschine",
                            "Übrige Kraftfahrzeuge",
                            "Fahrrad ohne Hilfsmotor",
                            "Pedelecs",
                            "Andere Fahrzeuge",
                            "Fußgänger",
                            "Andere Personen"
                        ],
                        "type": "string",
                        "description": "Filter nach Art der Verkehrsbeteiligung",
                        "name": "verkehrsart",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter nach Fehlverhalten",
                        "name": "fehlverhalten",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter nach Jahr",
                        "name": "jahr",
                        "in": "query"
                    },
                    {
                        "maximum": 12,
                        "minimum": 0,
                        "type": "integer",
                        "description": "Filter nach Monat (1-12, 0 für Ganzjahresdaten)",
                        "name": "monat",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.UnfallFehlverhalten"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid parameter type or range",
                        "schema": {
                            "$ref": "#/definitions/handlers.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Database execution or scanning failure",
                        "schema": {
                            "$ref": "#/definitions/handlers.HTTPError"
                        }
                    }
                }
            }
        },
        "/unfallFehlverhalten/jahre": {
            "get": {
                "description": "Gibt alle Jahre zurück, für die Daten vorhanden sind.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GENESIS-Online (Die Datenbank des Statistischen Bundesamtes)"
                ],
                "summary": "Verfügbare Jahre abrufen",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.YearsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Database execution or scanning failure",
                        "schema": {
                            "$ref": "#/definitions/handlers.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "data.UnfallBeteiligung": {
            "type": "object",
            "properties": {
                "altersgruppe": {
                    "type": "string"
                },
                "anzahl": {
                    "type": "integer"
                },
                "beteiligungsart": {
                    "description": "\"Unfallbeteiligte\" / \"Hauptverursacher des Unfalls\"",
                    "type": "string"
                },
                "geschlecht": {
                    "type": "string"
                },
                "jahr": {
                    "type": "integer"
                },
                "monat": {
                    "description": "1-12 for months, 0 for full year data",
                    "type": "integer"
                },
                "ortslage": {
                    "type": "string"
                },
                "unfallkategorie": {
                    "type": "string"
                },
                "verkehrsart": {
                    "type": "string"
                }
            }
        },
        "data.UnfallFehlverhalten": {
            "type": "object",
            "properties": {
                "anzahl": {
                    "type": "integer"
                },
                "fehlverhalten": {
                    "type": "string"
                },
                "jahr": {
                    "type": "integer"
                },
                "monat": {
                    "description": "1-12 for months, 0 for full year data",
                    "type": "integer"
                },
                "verkehrsart": {
                    "type": "string"
                }
            }
        },
...
        "handlers.HTTPError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid parameter format"
                }
            }
        },
        "handlers.YearsResponse": {
            "type": "object",
            "properties": {
                "jahre": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        2022
                    ]
                }
            }
        }
    }
}